add finalizers to allow deep delete of VMs
In this PR we added a feature of finalizier
which allow us a more safe way of deleting VMs
and releasing their mac.

The finalizers do not
allow the object to be deleted until
some necessary actions occur. in our
case- the release of the mac address.

Before this PR, the release of the mac
and the deletion of the VM were not
dependent,
so in order to keep track
with the allocated mac addresses
and safely release them,
we mapped each mac address to the
name of the object it was allocated for
(pod or VM).
that implementation could lead to
safety issues (for example, if
the controller collapses for some reason,
we lose track of
the allocated macs.

With the finalizers, we can be sure that
before any deletion of a VM
it's mac will be released, and tracking
the alocated mac addresses is
done through the VM itself.